Sensor fution gaming

+++ Data_parsing/BaselineAcc.hpp
@@ -20,7 +20,7 @@ public:
         this->variableNames_ = extraction.GetColumnNames();
 
         for (int count = 0; count < variableNames_.size(); count++) {
-            this->data_.push_back(extraction.GetColumn<float>(count));
+            this->data_.push_back(extraction.GetColumn<double>(count));
         }
 
         this->t_ = data_[getVariableNameIndex("t")];
@@ -38,53 +38,53 @@ public:
         throw std::runtime_error("Variable name not found: " + name);
     }
 
-    std::vector<float> getAccelerationVector(const int &index) {
+    std::vector<double> getAccelerationVector(const int &index) {
         return {ax_[index], ay_[index], az_[index]};
     }
 
-    std::vector<float> getAccelerationWithTime(const int &index) {
+    std::vector<double> getAccelerationWithTime(const int &index) {
         return {t_[index], ax_[index], ay_[index], az_[index]};
     }
 
-    std::vector<std::vector<float>> getAccelerationVectorColumn() {
-        std::vector<std::vector<float>> column;
+    std::vector<std::vector<double>> getAccelerationVectorColumn() {
+        std::vector<std::vector<double>> column;
         for (int i = 0; i < ax_.size(); i++) {
             column.push_back(getAccelerationVector(i));
         }
         return column;
     }
 
-    std::vector<std::vector<float>> getAccelerationVectorWithTimeColumn() {
-        std::vector<std::vector<float>> column;
+    std::vector<std::vector<double>> getAccelerationVectorWithTimeColumn() {
+        std::vector<std::vector<double>> column;
         for (int i = 0; i < ax_.size(); i++) {
             column.push_back(getAccelerationWithTime(i));
         }
         return column;
     }
 
-    std::vector<float> getSingleTypeColumn_t() const {
+    std::vector<double> getSingleTypeColumn_t() const {
         return t_;
     }
 
-    std::vector<float> getSingleTypeColumn_ax() const {
+    std::vector<double> getSingleTypeColumn_ax() const {
         return ax_;
     }
 
-    std::vector<float> getSingleTypeColumn_ay() const {
+    std::vector<double> getSingleTypeColumn_ay() const {
         return ay_;
     }
 
-    std::vector<float> getSingleTypeColumn_az() const {
+    std::vector<double> getSingleTypeColumn_az() const {
         return az_;
     }
 
 private:
-    std::vector<float> t_;
-    std::vector<float> ax_;
-    std::vector<float> ay_;
-    std::vector<float> az_;
+    std::vector<double> t_;
+    std::vector<double> ax_;
+    std::vector<double> ay_;
+    std::vector<double> az_;
     std::vector<std::string> variableNames_;
-    std::vector<std::vector<float>> data_;
+    std::vector<std::vector<double>> data_;
 };
 
 #endif//AIS4104_ASSIGNMENTS_BASELINEACC_HPP
